refactor(#9) :: assets 폴더 정리
