+++ src/assets/onboarding/team-introduce/Ara.jpeg
