+++ src/assets/onboarding/team-introduce/kangsiwoo.png
