+++ src/components/header/index.jsx
@@ -1,6 +1,6 @@
 import * as S from "./style.jsx";
-import Logo from '../../assets/Logo.svg';
-import Hambuger from '../../assets/hambuger.svg'
+import Logo from '../../assets/header/Logo.svg';
+import Hambuger from '../../assets/header/hambuger.svg'
 import {useState} from "react";
 import HeaderModal from '../modal/header'
