+++ src/components/onboarding/first/index.jsx
@@ -1,7 +1,7 @@
 import * as S from "./style.jsx";
 import Header from '../../header/index.jsx'
-import Logo from '../../../assets/Logo.svg';
-import Icon from '../../../assets/main.svg'
+import Logo from '../../../assets/header/Logo.svg';
+import Icon from '../../../assets/onboarding/first/main.svg'
 
 export default function Index(){
     return(
