+++ src/components/onboarding/team-introduce/index.jsx
@@ -1,15 +1,15 @@
 import * as S from './style.jsx';
 import {useState} from "react";
-import Dohun from '../../../assets/dohun.jpg';
-import Hyojun from '../../../assets/123isi.jpg';
-import Taeyoung from '../../../assets/noahmik.webp';
-import Jemin from '../../../assets/jmj732.png';
-import Heodongun from '../../../assets/heodongun.svg';
-import Siwoo from '../../../assets/kangsiwoo.png';
-import Soeun from '../../../assets/soeun823.jpeg'
-import Ara from '../../../assets/Ara.jpeg';
-import Yunchan from '../../../assets/yunchan.jpeg';
-import Huhon from '../../../assets/huhon.jpeg';
+import Dohun from '../../../assets/onboarding/team-introduce/dohun.jpg';
+import Hyojun from '../../../assets/onboarding/team-introduce/123isi.jpg';
+import Taeyoung from '../../../assets/onboarding/team-introduce/noahmik.webp';
+import Jemin from '../../../assets/onboarding/team-introduce/jmj732.png';
+import Heodongun from '../../../assets/onboarding/team-introduce/heodongun.svg';
+import Siwoo from '../../../assets/onboarding/team-introduce/kangsiwoo.png';
+import Soeun from '../../../assets/onboarding/team-introduce/soeun823.jpeg'
+import Ara from '../../../assets/onboarding/team-introduce/Ara.jpeg';
+import Yunchan from '../../../assets/onboarding/team-introduce/yunchan.jpeg';
+import Huhon from '../../../assets/onboarding/team-introduce/huhon.jpeg';
 
 export default function TeamIntroduce(){
     const [isGrade, setIsGrade] = useState(false);
